Other/sihModel/requirements.txt
torch>=2.0.0
torchvision
numpy
pandas
albumentations
opencv-python
Pillow
matplotlib
tqdm
scikit-learn
